requirements.txt
aiohttp==3.7.3
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
pip==20.3.3
pycparser==2.20
pylint==2.6.0
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
setuptools==51.0.0
six==1.15.0
slackclient==2.9.3
soupsieve==2.1
toml==0.10.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
wheel==0.36.2
wrapt==1.11.2
yarl==1.6.3
